requirements.txt
click
flask
requests
ruamel.yaml
semantic_version
waitress
